uppsrc: Fixing GCC 10.2. warnings
git-svn-id: svn://ultimatepp.org/upp/trunk@15182 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Fn.h
@@ -146,7 +146,7 @@ C gather(Args&& ...args)
 template <class C, typename... Args>
 int scatter_n(int n, const C& c, Args& ...args)
 {
-	if(n < sizeof...(args))
+	if(n < (int)sizeof...(args))
 		return 0;
 	iter_get(c.begin(), args...);
 	return sizeof...(args);
